extract set_identity
