+++ validator/src/commands/mod.rs
@@ -3,5 +3,6 @@ pub mod contact_info;
 pub mod exit;
 pub mod monitor;
 pub mod plugin;
+pub mod set_identity;
 pub mod staked_nodes_overrides;
 pub mod wait_for_restart_window;
